requirements.txt
langchain
langchain-community
langchain-mistralai
langchain-groq
python-dotenv
streamlit
streamlit-chat
bs4
faiss-cpu
sentence-transformers